Add files via upload
Here is an adapted version of the codes, included with an example c3d file. This file is just an example, with less markers than shown in the attached pdf. Unfortunatly, the original 3cd files used for the data presented in the article cannot be shared openly. if interested please contact the Laboratory for Movement Biomechanics at the ETH Zurich.